requirements.txt
torch>=2.2.0
transformers>=4.43.0
timm>=0.9.16
pillow>=10.0.0
matplotlib>=3.8.0
numpy>=1.24.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
safetensors>=0.4.0
